requirements.txt
chalice
google-analytics-data
mysql-connector-python
